+++ R/postcode_data_join.R
@@ -1,16 +1,26 @@
 #' Use postcodes.io to get postcode data
 #'
-#' @param df a data frame with a column of postcodes
-#' @param var the name of the variable in the data frame that contains postcodes (only!)
-#' @param fix_invalid whether to try to fix any postcodes that are not found (potentially because they are terminated codes)
+#' @param df A data frame with a column of postcodes.
+#' @param var The name of the variable in the data frame that contains
+#'   postcodes (only!). Should be acceptable as a symbol (variable name)
+#'   or as a standard character string.
+#' @param fix_invalid Whether to try to fix any postcodes that are not
+#'   found (potentially because they are terminated codes, or typos).
+#' @param narrow Whether to narrow the returned data frame by removing
+#'   a bunch of potentially unnecessary fields from the data before
+#'   returning the final product to the user. Boolean, default TRUE.
+#'   Set to FALSE to get all data variables available from the API,
+#'   plus a data quality indicator and a record of which codes were
+#'   invalid and subsequently fixed, if any.
 #'
 #' @importFrom assertthat assert_that
 #' @export
-postcode_data_join <- function(df, var = postcode, fix_invalid = TRUE) {
+postcode_data_join <- function(df, var = postcode, fix_invalid = TRUE, narrow = TRUE) {
 
   valid_results <- NULL
   terminated_results <- NULL
   autocompleted_results <- NULL
+  replaced <- c()
 
 
   # check that var is a column of df
@@ -46,44 +56,39 @@ postcode_data_join <- function(df, var = postcode, fix_invalid = TRUE) {
   # then if that fails, returning a nearby code using the autocomplete feature
   # (assuming that codes with only single final character different are
   # geographically near each other - I think this is sound).
-  if (fix_invalid & nrow(invalid) > 0) {
 
-    # filter out any invalid codes that match a terminated code...
-    terminated <- invalid %>%
-      dplyr::mutate(response = purrr::map(.data[["query_code"]], check_term_possibly)) %>%
-      dplyr::filter(!purrr::map_lgl(.data[["response"]], is.null))
+  if (nrow(invalid) > 0) {
+    if (fix_invalid) {
 
-    # ...and find the current code for the same lon/lat
-    # (only proceed with trying to expand the API response if there are
-    # matched codes!)
-    if (nrow(invalid) > 0) {
-      if (fix_invalid) {
+      # filter out any invalid codes that match a terminated code...
+      terminated <- invalid %>%
+        dplyr::mutate(response = purrr::map(.data[["query_code"]], check_term_possibly)) %>%
+        dplyr::filter(!purrr::map_lgl(.data[["response"]], is.null))
 
-        # filter out any invalid codes that match a terminated code...
-        terminated <- invalid %>%
-          dplyr::mutate(response = purrr::map(.data[["query_code"]], check_term_possibly)) %>%
-          dplyr::filter(!purrr::map_lgl(.data[["response"]], is.null))
+      # ...and find the current code for the same lon/lat
+      # (only proceed with trying to expand the API response if there are
+      # matched codes!)
+      if (nrow(terminated) > 0) {
+        lonlats <- terminated %>%
+          extract_lonlat()
+        # keep original codes as 'postcode' column
+        terminated_codes <- terminated %>%
+          dplyr::select(postcode = .data[["query_code"]])
 
-        if (nrow(terminated) > 0) {
-          lonlats <- terminated %>%
-            extract_lonlat()
-          # keep original codes as 'postcode' column
-          terminated_codes <- terminated %>%
-            dplyr::select(postcode = .data[["query_code"]])
+        terminated_results <- bulk_reverse_geocode(lonlats) %>%
+          dplyr::rename(query_code = .data[["postcode"]]) %>%
+          dplyr::bind_cols(terminated_codes, .) %>%
+          dplyr::select(!.data[["distance"]])
 
-          terminated_results <- bulk_reverse_geocode(lonlats) %>%
-            dplyr::rename(query_code = .data[["postcode"]]) %>%
-            dplyr::bind_cols(terminated_codes, .) %>%
-            dplyr::select(!.data[["distance"]])
-        }
-
-        usethis::ui_info(stringr::str_glue(
-          "The following postcodes are terminated:
-        {terminated_results$postcode}
-        and have been replaced with the following current postcodes:
-        {terminated_results$query_code}"
-        ))
+        replaced <- terminated_codes$postcode
       }
+
+      usethis::ui_info(stringr::str_glue(
+        "The following postcodes are terminated:
+          {terminated_results$postcode}
+          and have been replaced with the following current postcodes:
+          {terminated_results$query_code}"
+      ))
 
       # if there is a remainder of invalid codes that don't match terminated
       # codes, try the autocomplete method.
@@ -97,7 +102,7 @@ postcode_data_join <- function(df, var = postcode, fix_invalid = TRUE) {
           dplyr::mutate(query_code = purrr::map_chr(.data[["postcode"]], autocomplete_possibly)) %>%
           dplyr::filter(!purrr::map_lgl(.data[["query_code"]], is.null)) %>%
           # keep both original postcode and new postcode to query
-          dplyr::select(c(postcode, query_code))
+          dplyr::select(postcode, query_code)
 
         if (nrow(autocompleted) > 0) {
           autocompleted_results <- autocompleted[["query_code"]] %>%
@@ -105,8 +110,8 @@ postcode_data_join <- function(df, var = postcode, fix_invalid = TRUE) {
             purrr::map_df(bulk_lookup) %>%
             dplyr::left_join(autocompleted, ., by = c(query_code = "postcode"))
         }
+        replaced <- c(replaced, autocompleted$postcode)
       }
-      replaced <- c(terminated_codes$postcode, autocompleted$postcode)
       remainder <- setdiff(invalid$query_code, replaced)
 
       if (length(replaced) > 0) {
@@ -114,7 +119,7 @@ postcode_data_join <- function(df, var = postcode, fix_invalid = TRUE) {
           stringr::str_wrap(
             stringr::str_glue(
               "The following codes were found to be invalid, but were
-            successfully replaced with nearby valid codes: {remainder}"
+            successfully replaced with nearby valid codes: {replaced}"
             ), 72))
       }
 
@@ -149,6 +154,8 @@ postcode_data_join <- function(df, var = postcode, fix_invalid = TRUE) {
     .id = "result_type"
   ) %>%
     dplyr::relocate(query_code, .after = .data[["result_type"]])
+
+  if (narrow) postcode_data <- narrow(postcode_data)
 
   df %>%
     dplyr::left_join(postcode_data, by = vctrs::vec_c({{var}} := "postcode"))