+++ tests/testthat/test-commonest.R
@@ -1,5 +1,5 @@
 # set <- sample(letters, 20, replace = TRUE)
-set <- c("k", "g", "e", "z", "l", "n", "e", "e", "e", "y", "b", "h", "h", "s", "h", "u", "y", "y", "y", "a")
+set <- c("k", "g", "y", "z", "l", "n", "e", "e", "e", "y", "b", "h", "h", "s", "h", "u", "y", "y", "e", "a")
 
 set.seed(1234)
 
@@ -8,13 +8,16 @@ set.seed(1234)
     out <- vctrs::vec_count(set)
     testthat::expect_length(out, 2)
     testthat::expect_equal(nrow(out), length(unique(set)))
-    testthat::expect_identical(out$key[1], "y")
+    testthat::expect_identical(out$key[1], "e")
   })
 
-"test_overall" %>%
+"test_overall_y" %>%
   testthat::test_that({
-    testthat::expect_equal(commonest(set), "e")
-    testthat::expect_equal(commonest(set, first = FALSE), "y")
+    testthat::expect_equal(commonest(set, first = TRUE), "y")
+  })
+"test_overall_e" %>%
+  testthat::test_that({
+    testthat::expect_equal(commonest(set, first = FALSE), "e")
   })
 
 "test_summarise" %>%
@@ -27,7 +30,7 @@ set.seed(1234)
         dplyr::summarise(
           letter = commonest(letters)
         ),
-      structure(list(letter = "e"), class = c("tbl_df", "tbl", "data.frame"
+      structure(list(letter = "y"), class = c("tbl_df", "tbl", "data.frame"
       ), row.names = c(NA, -1L))
     )
     testthat::expect_equal(
@@ -35,7 +38,7 @@ set.seed(1234)
         dplyr::summarise(
           letter = commonest(letters, first = FALSE)
         ),
-      structure(list(letter = "y"), class = c("tbl_df", "tbl", "data.frame"
+      structure(list(letter = "e"), class = c("tbl_df", "tbl", "data.frame"
       ), row.names = c(NA, -1L))
     )
   })